viv-memory/viv-brain/requirements.txt
# Core dependencies
neo4j-python-driver>=5.15.0
mem0ai>=0.1.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Observability
langfuse>=2.0.0

# Scheduling & async
schedule>=1.2.0
asyncio>=3.4.3
aiohttp>=3.9.0

# Data processing
numpy>=1.24.0
scikit-learn>=1.3.0

# Time handling
python-dateutil>=2.8.0

# Logging
structlog>=23.0.0

# Graph processing
networkx>=3.0.0
